requirements.txt
Flask >= 3.0.0
Flask-WTF
PyYAML
redis
